Separate UpstreamEndpointInfo in a separate  file
