+++ include/envoy/upstream/BUILD
@@ -107,6 +107,7 @@ envoy_cc_library(
     deps = [
         "//include/envoy/config:typed_config_interface",
         "//include/envoy/upstream:types_interface",
+        "//include/envoy/protobuf:message_validator_interface",
         "//include/envoy/upstream:upstream_interface",
     ],
 )
