+++ source/extensions/filters/network/thrift_proxy/router/router_ratelimit_impl.cc
@@ -3,6 +3,7 @@
 #include "envoy/common/exception.h"
 #include "envoy/config/route/v3/route_components.pb.h"
 #include "envoy/ratelimit/ratelimit.h"
+#include "common/protobuf/utility.h"
 
 #include "extensions/filters/network/thrift_proxy/router/router.h"
